+++ x-pack/plugin/searchable-snapshots/src/main/java/org/elasticsearch/index/store/cache/FrozenIndexInput.java
@@ -12,20 +12,10 @@ import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.apache.lucene.codecs.CodecUtil;
 import org.apache.lucene.index.IndexFileNames;
-import org.apache.lucene.store.AlreadyClosedException;
 import org.apache.lucene.store.IOContext;
 import org.apache.lucene.store.IndexInput;
-import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.BytesRefIterator;
-import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.StepListener;
-import org.elasticsearch.blobstore.cache.BlobStoreCacheService;
-import org.elasticsearch.blobstore.cache.CachedBlob;
-import org.elasticsearch.common.bytes.BytesReference;
-import org.elasticsearch.common.lease.Releasable;
-import org.elasticsearch.common.unit.ByteSizeUnit;
 import org.elasticsearch.index.snapshots.blobstore.BlobStoreIndexShardSnapshot.FileInfo;
-import org.elasticsearch.index.store.BaseSearchableSnapshotIndexInput;
 import org.elasticsearch.index.store.IndexInputStats;
 import org.elasticsearch.index.store.SearchableSnapshotDirectory;
 import org.elasticsearch.xpack.searchablesnapshots.cache.ByteRange;
@@ -43,27 +33,11 @@ import java.util.function.Consumer;
 
 import static org.elasticsearch.xpack.searchablesnapshots.SearchableSnapshotsUtils.toIntBytes;
 
-public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
-
-    public static final IOContext CACHE_WARMING_CONTEXT = new IOContext();
+public class FrozenIndexInput extends MetadataCachingIndexInput {
 
     private static final Logger logger = LogManager.getLogger(FrozenIndexInput.class);
-    private static final int COPY_BUFFER_SIZE = ByteSizeUnit.KB.toIntBytes(8);
 
     private final FrozenCacheFile frozenCacheFile;
-    private final int defaultRangeSize;
-    private final int recoveryRangeSize;
-
-    /**
-     * If > 0, represents a logical file within a compound (CFS) file or is a slice thereof represents the offset of the logical
-     * compound file within the physical CFS file
-     */
-    private final long compoundFileOffset;
-
-    // last read position is kept around in order to detect (non)contiguous reads for stats
-    private long lastReadPosition;
-    // last seek position is kept around in order to detect forward/backward seeks for stats
-    private long lastSeekPosition;
 
     public FrozenIndexInput(
         String name,
@@ -83,13 +57,13 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
             0L,
             0L,
             fileInfo.length(),
+            new CacheFileReference(directory, fileInfo.physicalName(), fileInfo.length()),
             directory.getFrozenCacheFile(name, fileInfo.length()),
             rangeSize,
             recoveryRangeSize,
             directory.getBlobCacheByteRange(name, fileInfo.length()),
             ByteRange.EMPTY
         );
-        assert getBufferSize() <= BlobStoreCacheService.DEFAULT_CACHED_BLOB_SIZE; // must be able to cache at least one buffer's worth
         stats.incrementOpenCount();
     }
 
@@ -102,25 +76,30 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
         long offset,
         long compoundFileOffset,
         long length,
+        CacheFileReference cacheFileReference,
         FrozenCacheFile frozenCacheFile,
-        int rangeSize,
+        int defaultRangeSize,
         int recoveryRangeSize,
         ByteRange headerBlobCacheByteRange,
         ByteRange footerBlobCacheByteRange
     ) {
-        super(logger, name, directory, fileInfo, context, stats, offset, length, headerBlobCacheByteRange, footerBlobCacheByteRange);
+        super(
+            logger,
+            name,
+            directory,
+            fileInfo,
+            context,
+            stats,
+            offset,
+            compoundFileOffset,
+            length,
+            cacheFileReference,
+            defaultRangeSize,
+            recoveryRangeSize,
+            headerBlobCacheByteRange,
+            footerBlobCacheByteRange
+        );
         this.frozenCacheFile = frozenCacheFile;
-        this.lastReadPosition = this.offset;
-        this.lastSeekPosition = this.offset;
-        this.defaultRangeSize = rangeSize;
-        this.recoveryRangeSize = recoveryRangeSize;
-        this.compoundFileOffset = compoundFileOffset;
-        assert offset >= compoundFileOffset;
-    }
-
-    @Override
-    public void doClose() {
-        // nothing needed to be done here
     }
 
     private long getDefaultRangeSize() {
@@ -135,15 +114,14 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
     }
 
     @Override
-    protected void doReadInternal(ByteBuffer b) throws IOException {
-        ensureContext(ctx -> ctx != CACHE_WARMING_CONTEXT);
+    protected void doReadNonMetadata(ByteBuffer b) throws Exception {
         final long position = getAbsolutePosition() - compoundFileOffset;
         final int length = b.remaining();
 
         final ReentrantReadWriteLock luceneByteBufLock = new ReentrantReadWriteLock();
         final AtomicBoolean stopAsyncReads = new AtomicBoolean();
         // Runnable that, when called, ensures that async callbacks (such as those used by readCacheFile) are not
-        // accessing the byte buffer anymore that was passed to readInternal
+        // accessing the byte buffer anymore that was passed to doReadNonMetadataInternal
         // In particular, it's important to call this method before adapting the ByteBuffer's offset
         final Runnable preventAsyncBufferChanges = () -> {
             luceneByteBufLock.writeLock().lock();
@@ -154,113 +132,13 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
             }
         };
 
-        final String fileName = frozenCacheFile.getCacheKey().getFileName();
         logger.trace("readInternal: read [{}-{}] ([{}] bytes) from [{}]", position, position + length, length, this);
 
         try {
-            // Can we serve the read directly from disk? If so, do so and don't worry about anything else.
-
-            final StepListener<Integer> waitingForRead = frozenCacheFile.readIfAvailableOrPending(
-                ByteRange.of(position, position + length),
-                (channel, pos, relativePos, len) -> {
-                    final int read = readCacheFile(channel, pos, relativePos, len, b, position, true, luceneByteBufLock, stopAsyncReads);
-                    assert read <= length : read + " vs " + length;
-                    return read;
-                }
-            );
-
-            if (waitingForRead != null) {
-                final Integer read = waitingForRead.asFuture().get();
-                assert read == length;
-                assert luceneByteBufLock.getReadHoldCount() == 0;
-                preventAsyncBufferChanges.run();
-                b.position(read); // mark all bytes as accounted for
-                readComplete(position, length);
-                return;
-            }
-
-            // Requested data is not on disk, so try the cache index next.
-            final ByteRange indexCacheMiss; // null if not a miss
-
-            final ByteRange blobCacheByteRange = maybeReadFromBlobCache(position, length);
-            if (blobCacheByteRange != ByteRange.EMPTY) {
-                final CachedBlob cachedBlob = directory.getCachedBlob(fileName, blobCacheByteRange);
-                assert cachedBlob == CachedBlob.CACHE_MISS || cachedBlob == CachedBlob.CACHE_NOT_READY || cachedBlob.from() <= position;
-                assert cachedBlob == CachedBlob.CACHE_MISS || cachedBlob == CachedBlob.CACHE_NOT_READY || length <= cachedBlob.length();
-
-                if (cachedBlob == CachedBlob.CACHE_MISS || cachedBlob == CachedBlob.CACHE_NOT_READY) {
-                    // We would have liked to find a cached entry but we did not find anything: the cache on the disk will be requested
-                    // so we compute the region of the file we would like to have the next time. The region is expressed as a tuple of
-                    // {start, end} where positions are relative to the whole file.
-                    indexCacheMiss = blobCacheByteRange;
-
-                    // We must fill in a cache miss even if CACHE_NOT_READY since the cache index is only created on the first put.
-                    // TODO TBD use a different trigger for creating the cache index and avoid a put in the CACHE_NOT_READY case.
-                } else {
-                    logger.trace("reading [{}] bytes of file [{}] at position [{}] using cache index", length, fileName, position);
-                    stats.addIndexCacheBytesRead(cachedBlob.length());
-
-                    preventAsyncBufferChanges.run();
-                    try {
-                        final int sliceOffset = toIntBytes(position - cachedBlob.from());
-                        final BytesRefIterator cachedBytesIterator = cachedBlob.bytes().slice(sliceOffset, length).iterator();
-                        int copiedBytes = 0;
-                        BytesRef bytesRef;
-                        while ((bytesRef = cachedBytesIterator.next()) != null) {
-                            b.put(bytesRef.bytes, bytesRef.offset, bytesRef.length);
-                            copiedBytes += bytesRef.length;
-                        }
-                        assert copiedBytes == length : "copied " + copiedBytes + " but expected " + length;
-
-                        final ByteRange cachedRange = ByteRange.of(cachedBlob.from(), cachedBlob.to());
-                        frozenCacheFile.populateAndRead(
-                            cachedRange,
-                            cachedRange,
-                            (channel, channelPos, relativePos, len) -> Math.toIntExact(len),
-                            (channel, channelPos, relativePos, len, progressUpdater) -> {
-                                assert len <= cachedBlob.to() - cachedBlob.from();
-                                final long startTimeNanos = stats.currentTimeNanos();
-                                writeCacheFile(
-                                    channel,
-                                    cachedBlob.bytes().slice(toIntBytes(relativePos), toIntBytes(len)).streamInput(),
-                                    channelPos,
-                                    relativePos,
-                                    len,
-                                    progressUpdater,
-                                    startTimeNanos
-                                );
-                            },
-                            directory.cacheFetchAsyncExecutor()
-                        );
-                    } catch (Exception e) {
-                        logger.debug(
-                            new ParameterizedMessage(
-                                "failed to store bytes [{}-{}] of file [{}] obtained from index cache",
-                                cachedBlob.from(),
-                                cachedBlob.to(),
-                                fileName
-                            ),
-                            e
-                        );
-                        // oh well, no big deal, at least we can return them to the caller.
-                    }
-
-                    readComplete(position, length);
-
-                    return;
-                }
-            } else {
-                // requested range is not eligible for caching
-                indexCacheMiss = null;
-            }
-
-            // Requested data is also not in the cache index, so we must visit the blob store to satisfy both the target range and any
-            // miss in the cache index.
-
             final ByteRange startRangeToWrite = computeRange(position);
             final ByteRange endRangeToWrite = computeRange(position + length - 1);
             assert startRangeToWrite.end() <= endRangeToWrite.end() : startRangeToWrite + " vs " + endRangeToWrite;
-            final ByteRange rangeToWrite = startRangeToWrite.minEnvelope(endRangeToWrite).minEnvelope(indexCacheMiss);
+            final ByteRange rangeToWrite = startRangeToWrite.minEnvelope(endRangeToWrite);
 
             assert rangeToWrite.start() <= position && position + length <= rangeToWrite.end() : "["
                 + position
@@ -295,154 +173,15 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
                 directory.cacheFetchAsyncExecutor()
             );
 
-            if (indexCacheMiss != null) {
-
-                fillIndexCache(fileName, indexCacheMiss);
-                if (compoundFileOffset > 0L
-                    && indexCacheMiss.equals(headerBlobCacheByteRange)
-                    && footerBlobCacheByteRange != ByteRange.EMPTY) {
-                    fillIndexCache(fileName, footerBlobCacheByteRange);
-                }
-            }
-
             final int bytesRead = populateCacheFuture.asFuture().get();
             assert bytesRead == length : bytesRead + " vs " + length;
             assert luceneByteBufLock.getReadHoldCount() == 0;
 
             preventAsyncBufferChanges.run();
             b.position(bytesRead); // mark all bytes as accounted for
-        } catch (final Exception e) {
+        } finally {
             preventAsyncBufferChanges.run();
-
-            // may have partially filled the buffer before the exception was thrown, so try and get the remainder directly.
-            final int alreadyRead = length - b.remaining();
-            final int bytesRead = readDirectlyIfAlreadyClosed(position + alreadyRead, b, e);
-            assert alreadyRead + bytesRead == length : alreadyRead + " + " + bytesRead + " vs " + length;
-
-            // In principle we could handle an index cache miss here too, ensuring that the direct read was large enough, but this is
-            // already a rare case caused by an overfull/undersized cache.
         }
-
-        readComplete(position, length);
-    }
-
-    private void fillIndexCache(String fileName, ByteRange indexCacheMiss) {
-        final Releasable onCacheFillComplete = stats.addIndexCacheFill();
-        final int indexCacheMissLength = toIntBytes(indexCacheMiss.length());
-
-        // We assume that we only cache small portions of blobs so that we do not need to:
-        // - use a BigArrays for allocation
-        // - use an intermediate copy buffer to read the file in sensibly-sized chunks
-        // - release the buffer once the indexing operation is complete
-        final ByteBuffer byteBuffer = ByteBuffer.allocate(indexCacheMissLength);
-
-        final StepListener<Integer> readListener = frozenCacheFile.readIfAvailableOrPending(
-            indexCacheMiss,
-            (channel, channelPos, relativePos, len) -> {
-                assert len <= indexCacheMissLength;
-
-                if (len == 0) {
-                    return 0;
-                }
-
-                // create slice that is positioned to read the given values
-                final ByteBuffer dup = byteBuffer.duplicate();
-                final int newPosition = dup.position() + Math.toIntExact(relativePos);
-                assert newPosition <= dup.limit() : "newpos " + newPosition + " limit " + dup.limit();
-                assert newPosition + len <= byteBuffer.limit();
-                dup.position(newPosition);
-                dup.limit(newPosition + Math.toIntExact(len));
-
-                final int read = channel.read(dup, channelPos);
-                if (read < 0) {
-                    throw new EOFException("read past EOF. pos [" + relativePos + "] length: [" + len + "]");
-                }
-                // NB use Channels.readFromFileChannelWithEofException not readCacheFile() to avoid counting this in the stats
-                assert read == len;
-                return read;
-            }
-        );
-
-        if (readListener == null) {
-            // Normally doesn't happen, we're already obtaining a range covering all cache misses above, but theoretically
-            // possible in the case that the real populateAndRead call already failed to obtain this range of the file. In that
-            // case, simply move on.
-            onCacheFillComplete.close();
-        } else {
-            readListener.whenComplete(read -> {
-                assert read == indexCacheMissLength;
-                byteBuffer.position(read); // mark all bytes as accounted for
-                byteBuffer.flip();
-                final BytesReference content = BytesReference.fromByteBuffer(byteBuffer);
-                directory.putCachedBlob(fileName, indexCacheMiss.start(), content, new ActionListener<>() {
-                    @Override
-                    public void onResponse(Void response) {
-                        onCacheFillComplete.close();
-                    }
-
-                    @Override
-                    public void onFailure(Exception e1) {
-                        onCacheFillComplete.close();
-                    }
-                });
-            }, e -> onCacheFillComplete.close());
-        }
-    }
-
-    private void readComplete(long position, int length) {
-        stats.incrementBytesRead(lastReadPosition, position, length);
-        lastReadPosition = position + length;
-        lastSeekPosition = lastReadPosition;
-    }
-
-    private int readDirectlyIfAlreadyClosed(long position, ByteBuffer b, Exception e) throws IOException {
-        if (e instanceof AlreadyClosedException || (e.getCause() != null && e.getCause() instanceof AlreadyClosedException)) {
-            try {
-                // cache file was evicted during the range fetching, read bytes directly from blob container
-                final long length = b.remaining();
-                final byte[] copyBuffer = new byte[toIntBytes(Math.min(COPY_BUFFER_SIZE, length))];
-                logger.trace(
-                    () -> new ParameterizedMessage(
-                        "direct reading of range [{}-{}] for cache file [{}]",
-                        position,
-                        position + length,
-                        frozenCacheFile
-                    )
-                );
-
-                int bytesCopied = 0;
-                final long startTimeNanos = stats.currentTimeNanos();
-                try (InputStream input = openInputStreamFromBlobStore(position + compoundFileOffset, length)) {
-                    long remaining = length;
-                    while (remaining > 0) {
-                        final int len = (remaining < copyBuffer.length) ? (int) remaining : copyBuffer.length;
-                        int bytesRead = input.read(copyBuffer, 0, len);
-                        if (bytesRead == -1) {
-                            throw new EOFException(
-                                String.format(
-                                    Locale.ROOT,
-                                    "unexpected EOF reading [%d-%d] ([%d] bytes remaining) from %s",
-                                    position,
-                                    position + length,
-                                    remaining,
-                                    frozenCacheFile
-                                )
-                            );
-                        }
-                        b.put(copyBuffer, 0, bytesRead);
-                        bytesCopied += bytesRead;
-                        remaining -= bytesRead;
-                        assert remaining == b.remaining() : remaining + " vs " + b.remaining();
-                    }
-                    final long endTimeNanos = stats.currentTimeNanos();
-                    stats.addDirectBytesRead(bytesCopied, endTimeNanos - startTimeNanos);
-                }
-                return bytesCopied;
-            } catch (Exception inner) {
-                e.addSuppressed(inner);
-            }
-        }
-        throw new IOException("failed to read data from cache", e);
     }
 
     private static int positionalWrite(SharedBytes.IO fc, long start, ByteBuffer byteBuffer) throws IOException {
@@ -654,9 +393,9 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
         if (sliceCompoundFile) {
             sliceCompoundFileOffset = this.offset + sliceOffset;
             sliceFrozenCacheFile = directory.getFrozenCacheFile(sliceName, sliceLength);
-            sliceHeaderByteRange = directory.getBlobCacheByteRange(sliceName, sliceLength);
+            sliceHeaderByteRange = directory.getBlobCacheByteRange(sliceName, sliceLength).shift(sliceCompoundFileOffset);
             if (sliceHeaderByteRange.length() < sliceLength) {
-                sliceFooterByteRange = ByteRange.of(sliceLength - CodecUtil.footerLength(), sliceLength);
+                sliceFooterByteRange = ByteRange.of(sliceLength - CodecUtil.footerLength(), sliceLength).shift(sliceCompoundFileOffset);
             } else {
                 sliceFooterByteRange = ByteRange.EMPTY;
             }
@@ -676,6 +415,7 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
             this.offset + sliceOffset,
             sliceCompoundFileOffset,
             sliceLength,
+            cacheFileReference,
             sliceFrozenCacheFile,
             defaultRangeSize,
             recoveryRangeSize,
@@ -686,11 +426,4 @@ public class FrozenIndexInput extends BaseSearchableSnapshotIndexInput {
         return slice;
     }
 
-    private static boolean assertCurrentThreadMayWriteCacheFile() {
-        final String threadName = Thread.currentThread().getName();
-        assert isCacheFetchAsyncThread(threadName) : "expected the current thread ["
-            + threadName
-            + "] to belong to the cache fetch async thread pool";
-        return true;
-    }
 }
